LevelSelector Completed
Level selector its working, just needs more levels, also difficulty manager its working
