+++ Assets/Art/Sprites/Enemies/Dañadas/SilencioCorchea.png.meta
@@ -1,10 +1,13 @@
 fileFormatVersion: 2
-guid: 4650427232eb75049bf5d0806e7b414c
+guid: f0a826f6449e49f4d877b51ca2d8011d
 TextureImporter:
   internalIDToNameTable:
   - first:
-      213: 3817821035928854734
-    second: Cuadrada_0
+      213: -7071498333518943591
+    second: SilencioCorchea_0
+  - first:
+      213: -7347787901175100466
+    second: SilencioCorchea_1
   externalObjects: {}
   serializedVersion: 13
   mipmaps:
@@ -113,23 +116,45 @@ TextureImporter:
     serializedVersion: 2
     sprites:
     - serializedVersion: 2
-      name: Cuadrada_0
+      name: SilencioCorchea_0
       rect:
         serializedVersion: 2
-        x: 670
-        y: 248
-        width: 580
+        x: 0
+        y: 34
+        width: 329
         height: 581
       alignment: 0
-      pivot: {x: 0.5, y: 0.5}
+      pivot: {x: 0, y: 0}
       border: {x: 0, y: 0, z: 0, w: 0}
       customData: 
       outline: []
       physicsShape: []
       tessellationDetail: -1
       bones: []
-      spriteID: ec0ad862100abf430800000000000000
-      internalID: 3817821035928854734
+      spriteID: 99a26d86bbdfcdd90800000000000000
+      internalID: -7071498333518943591
+      vertices: []
+      indices: 
+      edges: []
+      weights: []
+    - serializedVersion: 2
+      name: SilencioCorchea_1
+      rect:
+        serializedVersion: 2
+        x: 306
+        y: 590
+        width: 12
+        height: 11
+      alignment: 0
+      pivot: {x: 0, y: 0}
+      border: {x: 0, y: 0, z: 0, w: 0}
+      customData: 
+      outline: []
+      physicsShape: []
+      tessellationDetail: -1
+      bones: []
+      spriteID: ecb8b8985d9670a90800000000000000
+      internalID: -7347787901175100466
       vertices: []
       indices: 
       edges: []
@@ -138,7 +163,7 @@ TextureImporter:
     customData: 
     physicsShape: []
     bones: []
-    spriteID: 5e97eb03825dee720800000000000000
+    spriteID: 
     internalID: 0
     vertices: []
     indices: 
@@ -148,7 +173,8 @@ TextureImporter:
     spriteCustomMetadata:
       entries: []
     nameFileIdTable:
-      Cuadrada_0: 3817821035928854734
+      SilencioCorchea_0: -7071498333518943591
+      SilencioCorchea_1: -7347787901175100466
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
