+++ Assets/Art/Sprites/Enemies/Dañadas/SilencioRedonda.png.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: 4650427232eb75049bf5d0806e7b414c
+guid: 5faf211a5a3c58e4381293dde0aea8a4
 TextureImporter:
   internalIDToNameTable:
   - first:
-      213: 3817821035928854734
-    second: Cuadrada_0
+      213: 4524937408866310484
+    second: SilencioRedonda_0
   externalObjects: {}
   serializedVersion: 13
   mipmaps:
@@ -113,23 +113,23 @@ TextureImporter:
     serializedVersion: 2
     sprites:
     - serializedVersion: 2
-      name: Cuadrada_0
+      name: SilencioRedonda_0
       rect:
         serializedVersion: 2
-        x: 670
-        y: 248
-        width: 580
-        height: 581
+        x: 11
+        y: 3
+        width: 368
+        height: 101
       alignment: 0
-      pivot: {x: 0.5, y: 0.5}
+      pivot: {x: 0, y: 0}
       border: {x: 0, y: 0, z: 0, w: 0}
       customData: 
       outline: []
       physicsShape: []
       tessellationDetail: -1
       bones: []
-      spriteID: ec0ad862100abf430800000000000000
-      internalID: 3817821035928854734
+      spriteID: 4595a56699ecbce30800000000000000
+      internalID: 4524937408866310484
       vertices: []
       indices: 
       edges: []
@@ -138,7 +138,7 @@ TextureImporter:
     customData: 
     physicsShape: []
     bones: []
-    spriteID: 5e97eb03825dee720800000000000000
+    spriteID: 
     internalID: 0
     vertices: []
     indices: 
@@ -148,7 +148,7 @@ TextureImporter:
     spriteCustomMetadata:
       entries: []
     nameFileIdTable:
-      Cuadrada_0: 3817821035928854734
+      SilencioRedonda_0: 4524937408866310484
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
