+++ Assets/Art/Sprites/Enemies/Muertes/SilencioSemiCorchea.png.meta
@@ -1,10 +1,22 @@
 fileFormatVersion: 2
-guid: 4650427232eb75049bf5d0806e7b414c
+guid: 9b5e5567966bc764f9fbd2728775b153
 TextureImporter:
   internalIDToNameTable:
   - first:
-      213: 3817821035928854734
-    second: Cuadrada_0
+      213: 2243377942196373942
+    second: SilencioSemiCorchea_0
+  - first:
+      213: 1760113036525192442
+    second: SilencioSemiCorchea_1
+  - first:
+      213: 699916384151458176
+    second: SilencioSemiCorchea_2
+  - first:
+      213: -962007806855751660
+    second: SilencioSemiCorchea_3
+  - first:
+      213: -2747612749903926416
+    second: SilencioSemiCorchea_4
   externalObjects: {}
   serializedVersion: 13
   mipmaps:
@@ -113,23 +125,111 @@ TextureImporter:
     serializedVersion: 2
     sprites:
     - serializedVersion: 2
-      name: Cuadrada_0
+      name: SilencioSemiCorchea_0
       rect:
         serializedVersion: 2
-        x: 670
-        y: 248
-        width: 580
-        height: 581
+        x: 0
+        y: 0
+        width: 347
+        height: 226
       alignment: 0
-      pivot: {x: 0.5, y: 0.5}
+      pivot: {x: 0, y: 0}
       border: {x: 0, y: 0, z: 0, w: 0}
       customData: 
       outline: []
       physicsShape: []
       tessellationDetail: -1
       bones: []
-      spriteID: ec0ad862100abf430800000000000000
-      internalID: 3817821035928854734
+      spriteID: 6b9cf8d3a54122f10800000000000000
+      internalID: 2243377942196373942
+      vertices: []
+      indices: 
+      edges: []
+      weights: []
+    - serializedVersion: 2
+      name: SilencioSemiCorchea_1
+      rect:
+        serializedVersion: 2
+        x: 345
+        y: 0
+        width: 348
+        height: 226
+      alignment: 0
+      pivot: {x: 0, y: 0}
+      border: {x: 0, y: 0, z: 0, w: 0}
+      customData: 
+      outline: []
+      physicsShape: []
+      tessellationDetail: -1
+      bones: []
+      spriteID: af44e923b7d2d6810800000000000000
+      internalID: 1760113036525192442
+      vertices: []
+      indices: 
+      edges: []
+      weights: []
+    - serializedVersion: 2
+      name: SilencioSemiCorchea_2
+      rect:
+        serializedVersion: 2
+        x: 691
+        y: 0
+        width: 348
+        height: 226
+      alignment: 0
+      pivot: {x: 0, y: 0}
+      border: {x: 0, y: 0, z: 0, w: 0}
+      customData: 
+      outline: []
+      physicsShape: []
+      tessellationDetail: -1
+      bones: []
+      spriteID: 081d2d93e3a96b900800000000000000
+      internalID: 699916384151458176
+      vertices: []
+      indices: 
+      edges: []
+      weights: []
+    - serializedVersion: 2
+      name: SilencioSemiCorchea_3
+      rect:
+        serializedVersion: 2
+        x: 1037
+        y: 0
+        width: 348
+        height: 226
+      alignment: 0
+      pivot: {x: 0, y: 0}
+      border: {x: 0, y: 0, z: 0, w: 0}
+      customData: 
+      outline: []
+      physicsShape: []
+      tessellationDetail: -1
+      bones: []
+      spriteID: 418d5312ff246a2f0800000000000000
+      internalID: -962007806855751660
+      vertices: []
+      indices: 
+      edges: []
+      weights: []
+    - serializedVersion: 2
+      name: SilencioSemiCorchea_4
+      rect:
+        serializedVersion: 2
+        x: 1383
+        y: 0
+        width: 347
+        height: 226
+      alignment: 0
+      pivot: {x: 0, y: 0}
+      border: {x: 0, y: 0, z: 0, w: 0}
+      customData: 
+      outline: []
+      physicsShape: []
+      tessellationDetail: -1
+      bones: []
+      spriteID: 07b64ade2c48ed9d0800000000000000
+      internalID: -2747612749903926416
       vertices: []
       indices: 
       edges: []
@@ -138,7 +238,7 @@ TextureImporter:
     customData: 
     physicsShape: []
     bones: []
-    spriteID: 5e97eb03825dee720800000000000000
+    spriteID: 
     internalID: 0
     vertices: []
     indices: 
@@ -148,7 +248,11 @@ TextureImporter:
     spriteCustomMetadata:
       entries: []
     nameFileIdTable:
-      Cuadrada_0: 3817821035928854734
+      SilencioSemiCorchea_0: 2243377942196373942
+      SilencioSemiCorchea_1: 1760113036525192442
+      SilencioSemiCorchea_2: 699916384151458176
+      SilencioSemiCorchea_3: -962007806855751660
+      SilencioSemiCorchea_4: -2747612749903926416
   mipmapLimitGroupName: 
   pSDRemoveMatte: 0
   userData:
