+++ Assets/Resources/Prefabs/LevelSelectorPrefabs.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 72944321db6df1e48af43051908723f8
-PrefabImporter:
+guid: 903018d9c2f67a44c80687b980c3dd3f
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
