--- Assets/Src/UI/Menus/States/DificultySelector.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 504f87e2cd3d600418fe8ad42c04ade7
