+++ Assets/Src/UI/Menus/States/Main.cs
@@ -15,7 +15,7 @@ namespace UI.Menus.States
                 case EMenuButton.Play:
                     //TODO: Change to the name of Level selector (for beta the actual level)
                     //MenuManager.Instance.ChangeSceneAndState("BaseLevel", new Gameplay());
-                    MenuManager.Instance.SetState(new DificultySelector());
+                    MenuManager.Instance.SetState(new LevelSelector());
                     break;
                 case EMenuButton.Options:
                     MenuManager.Instance.SetState(new Options());
