--- Assets/TextMesh Pro/Fonts/SteampunkMachineryFont-Regular.ttf.meta
@@ -1,21 +1,0 @@
-fileFormatVersion: 2
-guid: 9c614a3432cde0a4e989d20c85fd4026
-TrueTypeFontImporter:
-  externalObjects: {}
-  serializedVersion: 4
-  fontSize: 16
-  forceTextureCase: -2
-  characterSpacing: 0
-  characterPadding: 1
-  includeFontData: 1
-  fontNames:
-  - Steampunk Machinery Font
-  fallbackFontReferences: []
-  customCharacters: 
-  fontRenderingMode: 0
-  ascentCalculationMode: 1
-  useLegacyBoundsCalculation: 0
-  shouldRoundAdvanceValue: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
