--- Assets/TextMesh Pro/Fonts/Teslafont-xpBO.ttf.meta
@@ -1,21 +1,0 @@
-fileFormatVersion: 2
-guid: 5190fcb7c0e3890418004b48e858bd10
-TrueTypeFontImporter:
-  externalObjects: {}
-  serializedVersion: 4
-  fontSize: 16
-  forceTextureCase: -2
-  characterSpacing: 0
-  characterPadding: 1
-  includeFontData: 1
-  fontNames:
-  - TESLAFONT
-  fallbackFontReferences: []
-  customCharacters: 
-  fontRenderingMode: 0
-  ascentCalculationMode: 1
-  useLegacyBoundsCalculation: 0
-  shouldRoundAdvanceValue: 1
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
